move ResourceAllocator out of the fg namespace
