+++ filament/src/fg/Blackboard.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "Blackboard.h"
+#include <fg/Blackboard.h>
 
 using namespace utils;
