+++ filament/src/fg/fg/RenderTargetResourceEntry.h
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-#ifndef TNT_RENDERTARGETRESOURCEENTRY_H
-#define TNT_RENDERTARGETRESOURCEENTRY_H
+#ifndef TNT_FILAMENT_FG_RENDERTARGETRESOURCEENTRY_H
+#define TNT_FILAMENT_FG_RENDERTARGETRESOURCEENTRY_H
 
 #include "fg/fg/ResourceEntry.h"
